feat(membership): add "none" option for expiration date and improve date calculation
Add a "none" option to the expiration date dropdown to allow for memberships without an expiration date. Improve the expiration date calculation logic by subtracting one day from the calculated date for "1year" and "3year" options. Also, add a display for the calculated expiration date in the UI to provide immediate feedback to the user.